adaptation Carousel to Mobile Responsive
