+++ packages/carousel/src/Carousel.stories.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Carousel } from '.'
 import { Box } from '@oneloop/box'
+import { Button } from '@oneloop/button'
 
 export default {
   component: Carousel,
@@ -11,8 +12,10 @@ export const CarouselComponent = () => {
   const imgs = [
     'https://www.wallpaperuse.com/wallp/0-9852_m.jpg',
     'https://static.inmofactory.com/images/inmofactory/documents/1/124683/34021431/582525589.jpg?rule=web_412x257',
-    'https://staticbp.com/img/prop_new_b/534/00534436-01.jpg',
     'https://imgs.nestimg.com/casa_en_fraccionamiento_privadas_de_la_hacienda_3210125690712336672.jpg',
+    'https://static.inmofactory.com/images/inmofactory/documents/1/124683/34021431/582525589.jpg?rule=web_412x257',
+    'https://imgs.nestimg.com/casa_en_fraccionamiento_privadas_de_la_hacienda_3210125690712336672.jpg',
+    'https://static.tokkobroker.com/pictures/3576822133737522479342344804526371670339487603874495331054752417644934016466.jpg',
   ]
 
   const VIDEO = [
@@ -24,6 +27,16 @@ export const CarouselComponent = () => {
 
   const Video360 = ['https://my.matterport.com/show/?m=eURj8Qwwzb2']
 
+  const OtherButton = () => (
+    <Button
+      onClick={() => alert('seleccionando Imágenes')}
+      variant={['primary', 'medium']}
+      padding="7px 16px"
+      borderRadius="8px"
+    >
+      {'Seleccionar Imágenes'}
+    </Button>
+  )
   return (
     <Box>
       <Carousel
@@ -32,7 +45,7 @@ export const CarouselComponent = () => {
         video={VIDEO}
         planos={PLANO}
         video360={Video360}
-        //height="326px"
+        //otherButton={<OtherButton />}
       />
     </Box>
   )
