+++ packages/carousel/src/components/ButtonGallery.js
@@ -1,0 +1,31 @@
+import { Box } from '@oneloop/box'
+import theme from '@oneloop/theme'
+
+export default function ButtonGallery({ text, ...props }) {
+  return (
+    <Box
+      __css={{
+        padding: '6px 12px',
+        color: '#FFF',
+        borderRadius: '32px',
+        background: theme.colors.secondary,
+        border: '1px solid #A2D4E5',
+        fontFamily: 'Nunito Sans',
+        fontSize: '12px',
+        fontWeight: '600',
+        cursor: 'pointer',
+        ':hover': {
+          background: '#A2D4E5',
+          color: theme.colors.secondary,
+        },
+        ':active': {
+          color: theme.colors.secondaryHover,
+          borderColor: theme.colors.secondary,
+        },
+      }}
+      {...props}
+    >
+      {text}
+    </Box>
+  )
+}
