+++ packages/carousel/styles/gallery.css
@@ -1,3 +1,4 @@
+@import '../../theme/styles/globals.css';
 .fullscreen {
   position: fixed;
   display: flex;
@@ -7,8 +8,7 @@
   width: 100%;
   height: 100%;
   flex-direction: column;
-  padding: 40px;
-  gap: 24px;
+  padding: 28px;
   overflow: hidden;
   top: 0;
   left: 0;
@@ -36,9 +36,11 @@
 .fsTabHeader {
   display: flex;
   padding: 8px 16px;
+  justify-content: space-between;
   align-items: center;
+  width: 100%;
   border-radius: 8px;
-  background: #fff;
+  background: var(--neutralGray8);
   gap: 20px;
 }
 
@@ -61,8 +63,9 @@
 }
 
 .fsContainerTabs {
-  width: 90%;
-  height: 90%;
+  padding-top: 30px;
+  width: 100%;
+  height: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -76,7 +79,7 @@
 }
 
 .fsContainerTabs i {
-  color: #fff;
+  color: var(--neutralGray8);
   cursor: pointer;
 }
 
@@ -86,26 +89,84 @@
   justify-content: center;
   align-items: center;
   height: 95%;
-  width: 95%;
+  width: 100%;
   gap: 24px;
 }
 
 .fsTabContImage img {
-  max-width: 95%;
-  max-height: 95%;
+  max-width: 100%;
+  max-height: 100%;
   height: auto;
 }
 
 .fsTabContImage p {
-  color: #fff;
+  color: var(--neutralGray8);
   position: absolute;
   bottom: 0px;
 }
 
 .fsCloseIcon {
   position: absolute;
-  color: #fff;
-  right: 4px;
-  top: 0px;
+  color: var(--neutralGray8);
+  right: 10px;
+  font-size: 16px;
+  top: 10px;
   cursor: pointer;
+}
+.iconPrev,
+.iconNext {
+  position: absolute;
+}
+.iconPrev {
+  left: -20px;
+}
+.iconNext {
+  right: -20px;
+}
+.iconPrevMobile,
+.iconNextMobile {
+  font-size: 24px;
+  color: var(--neutralGray8);
+}
+.buttonsMainImgContainer {
+  position: absolute;
+  bottom: 16px;
+  right: 16px;
+  display: flex;
+  gap: 8px;
+}
+body {
+  padding: 0 !important;
+}
+@media (min-width: 500px) {
+  .fullscreen {
+    padding: 40px;
+    gap: 24px;
+  }
+  .fsCloseIcon {
+    right: 4px;
+    font-size: 32px;
+    top: 0px;
+  }
+  .fsTabHeader {
+    padding: 8px 16px;
+    align-items: center;
+    width: auto;
+  }
+  .fsContainerTabs {
+    width: 90%;
+    padding-top: 0;
+  }
+  .iconPrev,
+  .iconNext {
+    position: relative;
+  }
+  .fsTabContImage {
+    width: 95%;
+  }
+  .fsTabContImage img {
+    max-width: 95%;
+    max-height: 95%;
+    height: auto;
+  }
 }
